# api/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger API PostgreSQL & Python - OpenAPI 3.0",
        "description": "Test",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.11"
    },
...
    "servers": [
        {
            "url": "http://192.168.0.107:5000/users"
        },
        {
            "url": "https://app-postgresql-python.vercel.app/users"
        }
    ],
...
    "paths": {
...
        "/get_userbyusername/{username}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get user by user name ✅",
                "description": "",
                "operationId": "username",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The name that needs to be fetched. Use PRE-1003938472 for testing. ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "id_user": {
                        "type": "integer",
                        "format": "int64",
                        "example": 10
                    },
                    "user_name": {
                        "type": "string",
                        "example": "PRE-100101010"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "James"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345"
                    },
                    "phone": {
                        "type": "string",
                        "example": "12345"
                    },
                    "date_born": {
                        "type": "string",
                        "example": "02/02/1995"
                    },
                    "register_date": {
                        "type": "string",
                        "example": "02/02/1995"
                    },
                    "address": {
                        "type": "string",
                        "description": "Address",
                        "format": "int32",
                        "example": "El Tejar - Ibarra"
                    },
                    "card_id_person": {
                        "type": "string",
                        "description": "Rol User",
                        "example": "1003938477"
                    },
                    "gender": {
                        "type": "integer",
                        "description": "Gender",
                        "format": "int32",
                        "example": 1
                    },
                    "id_rol": {
                        "type": "integer",
                        "description": "Rol User",
                        "format": "int32",
                        "example": 1
                    },
                    "user_state": {
                        "type": "integer",
                        "description": "User Status",
                        "format": "int32",
                        "example": 1
                    },
                    "user_delete": {
                        "type": "integer",
                        "description": "User Delete",
                        "format": "int32",
                        "example": 1
                    }
                },
                "xml": {
                    "name": "user"
                }
            },
...
        },
...
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}